recipe_app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
Flask-Login==0.6.3
bcrypt==4.1.3
pytest
pytest-cov